+++ src/components/TodoAddForm.tsx
@@ -1,42 +1,17 @@
-'use client';  
-//Attempted import error: 'useForm' is not exported from 'react-hook-form' (imported as 'useForm').
-//react-hook-form/issues/11284
-
-import { useRef } from 'react';
-import { useForm, SubmitHandler } from 'react-hook-form';
-import type { TodoStyle } from '@/type.ts';
+import { UseFormRegister , FieldErrors } from 'react-hook-form';
 
 type TodoAddFormProps = {
-  handleAddTodo: (data: TodoStyle) => void;
-};
+  onSubmit: () => void;
+  errors : FieldErrors<{title: string}>;
+  register: UseFormRegister<{title: string}>;
+}
 
-export const TodoAddForm = ({ handleAddTodo }: TodoAddFormProps) => {
-  //primary key. maybe uuid well be better
-  const indexRef = useRef(0);
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<{ title: string }>({
-    mode: 'onSubmit',
-    reValidateMode: 'onSubmit',
-  });
-
-  const onSubmit: SubmitHandler<{ title: string }> = (data) => {
-    handleAddTodo({ 
-      ...data, 
-      id: indexRef.current,
-      date: new Date().toISOString().split('T')[0] 
-    });
-    indexRef.current += 1;
-  };
+export const TodoAddForm = ({ onSubmit, errors,register }: TodoAddFormProps) => {
 
   return (
-    <form onSubmit={handleSubmit(onSubmit)}>
+    <form onSubmit={onSubmit}>
       <input {...register('title', { required: true })} />
       {errors.title && <span>🔺This field is required🔺</span>}
-
       <button type="submit">追加</button>
     </form>
   );
